+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [FEAT] Added `#finalized?` method to `Axn::Result` to check if result has completed execution
 * [FEAT] Added `type: :params` validation option for `expects`/`exposes` that accepts Hash or ActionController::Parameters (Rails-compatible)
 * [FEAT] Allow validations to access instance methods (e.g. `inclusion: { in: :some_method }`)
+* [FEAT] Allow message `prefix` to invoke callables/method name symbols the same way e.g. `if` does
 * [BREAKING] `default`s for `expects` and `exposes` are only applied if no value is passed in (previous applied if given value was blank, which caused bugs for boolean handling)
 
 ## 0.1.0-alpha.2.8.1